refactor: split docs data nav normalize modules
