+++ crates/ox_content_docs/src/normalize/tests.rs
@@ -1,0 +1,6 @@
+mod callables;
+mod entries;
+mod member_kinds;
+mod members;
+mod type_aliases;
+mod type_parameters;
